DEP-61 refactor : TokenGenerator에서 token resolve와 만료기한 검증 로직을 분리
